+++ include/minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dbejar-s <dbejar-s@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: pmarkaid <pmarkaid@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/01 15:11:45 by pmarkaid          #+#    #+#             */
-/*   Updated: 2024/09/02 12:58:17 by dbejar-s         ###   ########.fr       */
+/*   Updated: 2024/09/02 21:58:55 by pmarkaid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -101,16 +101,17 @@ typedef struct s_macro
 int				syntax_error_check(t_macro *macro, char *ins);
 
 /* tokenizer */
-int				tokenizer(t_macro *macro);
+int					tokenizer(t_macro *macro);
 t_list				*split_args_by_quotes(char *ins);
 
 /* tokenizer_expand */
 t_token				*expand_arg_tokens(t_macro *macro);
 void				ensure_at_least_one_cmd(t_token **tokens);
 
+/* tokenizer_retokenize */
+t_token				*handle_retokenize(char *expanded, t_token *token, t_macro *macro);
+
 /* tokenizer_utils */
-char				*expand_envir(char *clean, char *ins,
-						t_macro *macro);
 bool				is_builtin(t_token *token);
 bool				is_redir(t_token *token, char *redir_type);
 void				fix_redirections(char *ins);
@@ -127,7 +128,6 @@ void				cmd_add_back(t_cmd **cmds, t_cmd *new);
 t_cmd				*last_cmd(t_cmd *cmd);
 void				print_cmds(t_cmd *cmds);
 char				*enum_to_char(t_type type);
-bool				is_last_of_type(t_token *tokens, t_type type);
 int					tokens_size(t_token *tokens);
 void				remove_token(t_token **tokens, t_token *token);
 t_token				*remove_empty_envir_tokens(t_macro *macro);
@@ -179,8 +179,6 @@ int				dup_file_descriptors(t_macro *macro, t_cmd *cmd,
 						int read_end);
 
 /* clean utils*/
-char				*get_envir_name(char *str);
-char				*get_envir_value(char *str, t_macro *macro);
 bool				envir_must_be_expanded(char *ins, int index);
 bool				is_in_quote(char *str, int index);
