+++ src/tokenizer_expand.c
@@ -6,76 +6,13 @@
 /*   By: pmarkaid <pmarkaid@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/22 17:45:23 by pmarkaid          #+#    #+#             */
-/*   Updated: 2024/09/02 21:42:52 by pmarkaid         ###   ########.fr       */
+/*   Updated: 2024/09/02 21:57:13 by pmarkaid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-static t_token	*build_retokens(t_list *lexemes, t_type type)
-{
-	t_token	*token;
-	t_token	*retokens;
 
-	retokens = NULL;
-	while (lexemes)
-	{
-		token = init_token();
-		if (!token)
-			return (NULL);
-		if (type == CMD)
-		{
-			token->type = CMD;
-			type = ARG;
-		}
-		else
-			token->type = ARG;
-		token->value = ft_strdup(lexemes->content);
-		if (!token->value)
-			return (free_tokens(&token));
-		token_add_back(&retokens, token);
-		lexemes = lexemes->next;
-	}
-	return (retokens);
-}
-
-static t_token	*plug_retokens(t_token *token, t_token *retokens,
-		t_macro *macro)
-{
-	t_token	*next;
-	t_token	*prev;
-	t_token	*last;
-
-	next = token->next;
-	prev = macro->tokens;
-	while (prev && prev->next != token)
-		prev = prev->next;
-	if (prev)
-		prev->next = retokens;
-	else
-		macro->tokens = retokens;
-	last = last_token(retokens);
-	if (last)
-		last->next = next;
-	token->next = NULL;
-	free_tokens(&token);
-	return (last);
-}
-
-static t_token	*retokenize(char *expanded, t_token *token)
-{
-	t_list	*lexemes;
-	t_token	*retokens;
-
-	lexemes = split_args_by_quotes(expanded);
-	if (!lexemes)
-		return (NULL);
-	retokens = build_retokens(lexemes, token->type);
-	if (!retokens)
-		return (NULL);
-	ft_lstclear(&lexemes, ft_del);
-	return (retokens);
-}
 
 char	*handle_literal_string(t_token *token, char *expanded)
 {
@@ -88,17 +25,7 @@ char	*handle_literal_string(t_token *token, char *expanded)
 	return (copy);
 }
 
-t_token	*handle_retokenize(char *expanded, t_token *token, t_macro *macro)
-{
-	t_token	*retokens;
-	t_token	*last_retokens;
 
-	retokens = retokenize(expanded, token);
-	if (!retokens)
-		return (NULL);
-	last_retokens = plug_retokens(token, retokens, macro);
-	return (last_retokens);
-}
 
 char	*expand_token(t_token *token, t_macro *macro)
 {
